Add better error checking for setup.sh

+++ setup.sh
@@ -2,14 +2,24 @@
 set -e
 
 echo ""
-echo "------------------"
-echo "- Dotfiles setup -"
-echo "------------------"
-echo ""
-echo "WARNING: The dotfiles repository should be cloned at '~/.dotfiles',"
-echo "         so this script should be at '~/.dotfiles/setup.sh'"
-echo ""
+echo "--------------------"
+echo "-- Dotfiles setup --"
+echo "--------------------"
 
+if [ ! -x "$HOME/.dotfiles/setup.sh" ]; then
+  echo ""
+  echo "ERROR: The dotfiles repository should be cloned at '~/.dotfiles',"
+  echo "         so this script should be at '~/.dotfiles/setup.sh'"
+  exit
+fi
+
+if [ "$(pwd)" != "$HOME/.dotfiles" ]; then
+  echo ""
+  echo "ERROR: This script must be run from $HOME/.dotfiles"
+  exit
+fi
+
+echo ""
 echo "Symlinking config files..."
 
 rm ~/.gitconfig
